Move `isNotNull` into client/utils
Shared utils live in client.
