+++ packages/client/utils/predicates.ts
