+++ packages/server/graphql/mutations/helpers/pushEstimateToGitHub.ts
@@ -28,7 +28,7 @@ import getIssueId from '../../../utils/githubQueries/getIssueId.graphql'
 import getRepoLabels from '../../../utils/githubQueries/getRepoLabels.graphql'
 import removeLabels from '../../../utils/githubQueries/removeLabels.graphql'
 import makeScoreGitHubComment from '../../../utils/makeScoreGitHubComment'
-import {isNotNull} from '../../../utils/predicates'
+import {isNotNull} from 'parabol-client/utils/predicates'
 import {GQLContext} from '../../graphql'
 import {ITaskEstimateInput} from '../../types/TaskEstimateInput'
 import MeetingPoker from '../../../database/types/MeetingPoker'
